Pick up changes to /etc/resolv.conf, etc. on the fly
Override the default handling of /etc/resolv.conf and similar from podman
and replace it with the Flatpak method where /etc/resolv.conf is a symlink
into /run/host/monitor/resolv.conf, and /run/host/monitor is a directory
mounted from the host.

This means that if /etc/resolv.conf is atomically rewritten by creating
a new file and moving it over the old file, the changes show up in the
client - which wouldn't happen with a simple bind mount.

We reuse flatpak-session-helper's monitoring rather than creating our
own monitor process.